refactor: Split logic from action declaration
